Rename decimaljson to customjson, close #225.
