+++ ichnaea/service/error.py
@@ -4,7 +4,7 @@ from colander import Invalid
 from pyramid.httpexceptions import HTTPError
 from pyramid.response import Response
 
-from ichnaea.decimaljson import (
+from ichnaea.customjson import (
     dumps,
     loads,
 )
