+++ ichnaea/service/search/tests/test_views.py
@@ -1,7 +1,7 @@
 from sqlalchemy import text
 from webob.response import gzip_app_iter
 
-from ichnaea.decimaljson import dumps, loads
+from ichnaea.customjson import dumps, loads
 from ichnaea.heka_logging import RAVEN_ERROR
 from ichnaea.models import (
     ApiKey,
