+++ ichnaea/tasks.py
@@ -74,7 +74,7 @@ class DatabaseTask(Task):
 
             # We feed the task arguments through the de/serialization process
             # to make sure the arguments can indeed be serialized.
-            # It's easy enough to put decimal, datetime, set or other
+            # It's easy enough to put datetime, set or other
             # non-serializable objects into the task arguments
             task_args = isinstance(args, tuple) and args or tuple(args)
             serializer = self.app.conf.CELERY_TASK_SERIALIZER
